+++ SleepModel/UIFont+HEMStyle.h
@@ -26,6 +26,7 @@
 + (UIFont*)bodyBold;
 + (UIFont*)bodySmall;
 + (UIFont*)bodySmallBold;
++ (UIFont*)buttonLarge;
 + (UIFont*)button;
 + (UIFont*)buttonSmall;