+++ SleepModel/UIFont+HEMStyle.m
@@ -87,6 +87,10 @@ static NSString* const HEMNumberFontFamilyNameMedium = @"AvenirNext-Medium";
     return [UIFont fontWithName:HEMFontFamilyNameModernMedium size:14.0f];
 }
 
++ (UIFont*)buttonLarge {
+    return [UIFont fontWithName:HEMFontFamilyNameModernRegular size:24.0f];
+}
+
 + (UIFont*)button {
     return [UIFont fontWithName:HEMFontFamilyNameModernMedium size:14.0f];
 }